pyproject.toml
[tool.poetry]
name = "availability-checker"
version = "0.1.0"
description = "To develop a function returning the real-time availability of professionals, updating based on current activity or schedule, the requirements and findings are as follows:"
authors = ["AutoGPT <info@agpt.co>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11"
fastapi = "^0.75.0"
passlib = {version = "^1.7.4", extras = ["bcrypt"]}
prisma = "*"
pydantic = "*"
python-jose = {version = "^3.3.0", extras = ["cryptography"]}
uvicorn = "*"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
